requirements.txt
pycrypto==2.6.1
Flask==1.1.1
SQLAlchemy==1.2.19

